requirements.txt
ccxt==4.1.13
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
websockets==11.0.3
aiohttp==3.8.5
asyncio==3.4.3
python-binance==1.0.19
colorama==0.4.6
tabulate==0.9.0
